use Selenide Conditions instead of Assertions
ref: https://github.com/selenide/selenide/wiki/Selenide-vs-Selenium
