+++ medusa-showcase/src/test/resources/application.properties
@@ -1,0 +1,2 @@
+# integration tests: open browser
+#selenide.headless=false
